editoast: remove unused fixtures
